fix(app-headless-cms): fix dynamic zone template edit dialog and description overflow
- Lift TemplateDialog state from DynamicZoneTemplate to DynamicZone parent
  so the edit dialog renders outside the Accordion and opens regardless of
  whether the accordion item is collapsed
- Use defaultOpen instead of open on Accordion.Item to keep it uncontrolled
  after parent re-renders
- Remove truncate from TemplateListItem description to allow text wrapping
  in list view
- Add line-clamp-2 to TemplateGridItem description to limit grid card
  height while still showing meaningful text
