+++ packages/app-headless-cms/src/admin/plugins/fieldRenderers/dynamicZone/TemplateListItem.tsx
@@ -28,7 +28,7 @@ export const TemplateListItem = ({ template, onTemplate }: TemplateListItemProps
                         {template.name}
                     </Text>
                     {template.description && (
-                        <Text size={"sm"} as={"div"} className={"text-neutral-muted truncate"}>
+                        <Text size={"sm"} as={"div"} className={"text-neutral-muted"}>
                             {template.description}
                         </Text>
                     )}
